merged: Some changes from linuxport branch for easier merging (mostly just case sensitivity and code ordering)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@9644 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ guilib/GUIRSSControl.cpp
@@ -1,14 +1,14 @@
 #include "include.h"
 #include "GUIRSSControl.h"
 #include "GUIWindowManager.h"
-#include "..\xbmc\settings.h"
+#include "../xbmc/Settings.h"
 #include "../xbmc/utils/CriticalSection.h"
-#include "..\xbmc\utils\SingleLock.h"
+#include "../xbmc/utils/SingleLock.h"
 
 
 CGUIRSSControl::CGUIRSSControl(DWORD dwParentID, DWORD dwControlId, float posX, float posY, float width, float height, const CLabelInfo& labelInfo, D3DCOLOR dwChannelColor, D3DCOLOR dwHeadlineColor, CStdString& strRSSTags)
 : CGUIControl(dwParentID, dwControlId, posX, posY, width, height),
-  m_scrollInfo(-1,1.0f,"")
+  m_scrollInfo((DWORD) -1,1.0f,"")
 {
   m_label = labelInfo;
   m_dwChannelColor = dwChannelColor;
